chore: change balance type from integer to bigint

+++ apps/banking_api/priv/repo/migrations/20210414220740_create_accounts_table.exs
@@ -5,7 +5,7 @@ defmodule BankingApi.Repo.Migrations.CreateAccounts do
     create table(:accounts, primary_key: false) do
       add :id, :uuid, primary_key: true
       add :user, :string
-      add :balance, :integer
+      add :balance, :bigint
 
       timestamps()
     end
